move schedules_management MCP server to metadata pattern
